+++ txn/command.go
@@ -39,8 +39,8 @@ const (
 	UndelegateCommand Command = 0x4E
 	// CheckpointRestoreCommand ...
 	CheckpointRestoreCommand Command = 0x4F
-	// KeyRotateSubmissionCommand ...
-	KeyRotateSubmissionCommand Command = 0x50
+	// RotateKeySubmissionCommand ...
+	RotateKeySubmissionCommand Command = 0x50
 	// StateVariableProposalCommand ...
 	StateVariableProposalCommand Command = 0x51
 	// TransferFundsCommand ...
@@ -49,36 +49,39 @@ const (
 	CancelTransferFundsCommand Command = 0x55
 	// ValidatorHeartbeat ...
 	ValidatorHeartbeatCommand Command = 0x56
+	// RotateEthereumKeySubmissionCommand ...
+	RotateEthereumKeySubmissionCommand Command = 0x57
 )
 
 var commandName = map[Command]string{
-	SubmitOrderCommand:              "Submit Order",
-	CancelOrderCommand:              "Cancel Order",
-	AmendOrderCommand:               "Amend Order",
-	WithdrawCommand:                 "Withdraw",
-	ProposeCommand:                  "Proposal",
-	VoteCommand:                     "Vote on Proposal",
-	AnnounceNodeCommand:             "Register new Node",
-	NodeVoteCommand:                 "Node Vote",
-	NodeSignatureCommand:            "Node Signature",
-	LiquidityProvisionCommand:       "Liquidity Provision Order",
-	CancelLiquidityProvisionCommand: "Cancel LiquidityProvision Order",
-	AmendLiquidityProvisionCommand:  "Amend LiquidityProvision Order",
-	ChainEventCommand:               "Chain Event",
-	SubmitOracleDataCommand:         "Submit Oracle Data",
-	DelegateCommand:                 "Delegate",
-	UndelegateCommand:               "Undelegate",
-	CheckpointRestoreCommand:        "Checkpoint Restore",
-	KeyRotateSubmissionCommand:      "Key Rotate Submission",
-	StateVariableProposalCommand:    "State Variable Proposal",
-	TransferFundsCommand:            "Transfer Funds",
-	CancelTransferFundsCommand:      "Cancel Transfer Funds",
-	ValidatorHeartbeatCommand:       "Validator Heartbeat",
+	SubmitOrderCommand:                 "Submit Order",
+	CancelOrderCommand:                 "Cancel Order",
+	AmendOrderCommand:                  "Amend Order",
+	WithdrawCommand:                    "Withdraw",
+	ProposeCommand:                     "Proposal",
+	VoteCommand:                        "Vote on Proposal",
+	AnnounceNodeCommand:                "Register new Node",
+	NodeVoteCommand:                    "Node Vote",
+	NodeSignatureCommand:               "Node Signature",
+	LiquidityProvisionCommand:          "Liquidity Provision Order",
+	CancelLiquidityProvisionCommand:    "Cancel LiquidityProvision Order",
+	AmendLiquidityProvisionCommand:     "Amend LiquidityProvision Order",
+	ChainEventCommand:                  "Chain Event",
+	SubmitOracleDataCommand:            "Submit Oracle Data",
+	DelegateCommand:                    "Delegate",
+	UndelegateCommand:                  "Undelegate",
+	CheckpointRestoreCommand:           "Checkpoint Restore",
+	RotateKeySubmissionCommand:         "Key Rotate Submission",
+	StateVariableProposalCommand:       "State Variable Proposal",
+	TransferFundsCommand:               "Transfer Funds",
+	CancelTransferFundsCommand:         "Cancel Transfer Funds",
+	ValidatorHeartbeatCommand:          "Validator Heartbeat",
+	RotateEthereumKeySubmissionCommand: "Ethereum Key Rotate Submission",
 }
 
 func (cmd Command) IsValidatorCommand() bool {
 	switch cmd {
-	case CheckpointRestoreCommand, NodeSignatureCommand, ChainEventCommand, NodeVoteCommand, ValidatorHeartbeatCommand, KeyRotateSubmissionCommand, StateVariableProposalCommand:
+	case CheckpointRestoreCommand, NodeSignatureCommand, ChainEventCommand, NodeVoteCommand, ValidatorHeartbeatCommand, RotateKeySubmissionCommand, StateVariableProposalCommand, RotateEthereumKeySubmissionCommand:
 		return true
 	default:
 		return false