+++ validators/topology_snapshot_test.go
@@ -101,6 +101,22 @@ func TestTopologySnapshot(t *testing.T) {
 	err = top.AddKeyRotate(ctx, nr2.Id, 5, kr2)
 	assert.NoError(t, err)
 
+	ekr1 := &commandspb.EthereumKeyRotateSubmission{
+		TargetBlock:    10,
+		CurrentAddress: "eth-address",
+		NewAddress:     "0x69bA3B3e6B5b1226A2e26De9a9E2D9C98f2b144B",
+	}
+	err = top.RotateEthereumKey(ctx, nr1.Id, 5, ekr1)
+	assert.NoError(t, err)
+
+	ekr2 := &commandspb.EthereumKeyRotateSubmission{
+		TargetBlock:    11,
+		CurrentAddress: "eth-address-2",
+		NewAddress:     "0xd6B6e9514f2793Af89745Fd69FDa0DAbC228d336",
+	}
+	err = top.RotateEthereumKey(ctx, nr2.Id, 5, ekr2)
+	assert.NoError(t, err)
+
 	// Check the hashes have changed after each state change
 	h3, err := top.GetHash(topKey)
 	require.Nil(t, err)
@@ -126,8 +142,10 @@ func TestTopologySnapshot(t *testing.T) {
 	assert.ElementsMatch(t, top.AllNodeIDs(), snapTop.AllNodeIDs())
 	assert.ElementsMatch(t, top.AllVegaPubKeys(), snapTop.AllVegaPubKeys())
 	assert.Equal(t, top.IsValidator(), snapTop.IsValidator())
-	assert.Equal(t, top.GetPendingKeyRotation(kr1.TargetBlock, kr1.NewPubKey), snapTop.GetPendingKeyRotation(kr1.TargetBlock, kr1.NewPubKey))
-	assert.Equal(t, top.GetPendingKeyRotation(kr2.TargetBlock, kr2.NewPubKey), snapTop.GetPendingKeyRotation(kr2.TargetBlock, kr2.NewPubKey))
+	assert.Equal(t, top.GetPendingKeyRotation(kr1.TargetBlock, nr1.Id), snapTop.GetPendingKeyRotation(kr1.TargetBlock, nr1.Id))
+	assert.Equal(t, top.GetPendingKeyRotation(kr2.TargetBlock, nr2.Id), snapTop.GetPendingKeyRotation(kr2.TargetBlock, nr2.Id))
+	assert.Equal(t, top.GetPendingEthereumKeyRotation(ekr1.TargetBlock, nr1.Id), snapTop.GetPendingEthereumKeyRotation(ekr1.TargetBlock, nr1.Id))
+	assert.Equal(t, top.GetPendingEthereumKeyRotation(ekr2.TargetBlock, nr2.Id), snapTop.GetPendingEthereumKeyRotation(ekr2.TargetBlock, nr2.Id))
 }
 
 func updateValidatorPerformanceToNonDefaultState(t *testing.T, top *validators.Topology) {